Extension of SmartDuplicate.py
Love your idea with a smart line duplication, so I took it and ran with it. I use a dictionary of keywords to replace (so you can add and remove at will, even using regex's in the keys), and I think with a little creativity in the key regex we could probably solve the issue you posted about auto-incrementing numbers.

Hope you like it!